VoiceAssist/requirements.txt
annotated-types==0.7.0

anyio==4.6.2.post1

certifi==2024.8.30

openai=1.54.4

cffi==1.17.1

charset-normalizer==3.4.0

colorama==0.4.6

cryptography==43.0.3

distro==1.9.0

faiss-cpu==1.9.0

filelock==3.16.1

fsspec==2024.10.0

h11==0.14.0

httpcore==1.0.6

httpx==0.27.1

huggingface-hub==0.26.2

idna==3.10

jinja2==3.1.4

jiter==0.7.1

joblib==1.4.2

levenshtein==0.26.1

llvmlite==0.43.0

MarkupSafe==3.0.2

more-itertools==10.5.0

mpmath==1.3.0

networkx==3.4.2

numba==0.60.0

numpy==2.0.2

openai-whisper==20240930

packaging==24.2

pdfminer.six==20231228

pdfplumber==0.11.4

pillow==11.0.0

protobuf==5.28.3

PyAudio==0.2.14

pycparser==2.22

pydantic==2.9.2

pydantic_core==2.23.4

PyPDF2==3.0.1

pypdfium2==4.30.0

pypiwin32==223

pyttsx3==2.98

pywin32==308

RapidFuzz==3.10.1

regex==2024.11.6

requests==2.32.3

safetensors==0.4.5

scikit-learn==1.5.2

scipy==1.14.1

sentence-transformers==3.3.0

setuptools==75.5.0

sniffio==1.3.1

SpeechRecognition==3.11.0

sympy==1.13.1

threadpoolctl==3.5.0

tiktoken==0.8.0

torch==2.5.1

tqdm==4.67.0

transformers==4.46.2

typing_extensions==4.12.2

urllib3==2.2.2